Rename navbar files
